YUK-9102 Stars Res Cust - trying to 'Save and Send' a Schedule is causing the 'error has occurred' page to display
Fix for the "IndexOutOfBoundsException: Index: 0, Size: 0" error.  This is caused by a schedule in the database missing schedule entries that it should have for its selected mode.  It should not be possible for a schedule to be saved with this problem in the current code.  If such a schedule is loaded and saved again, it will now have the missing entries added back in, and will not cause this error.