YUK-9558 Stars operator & consumer - Schedule page, Recommended Settings doesn't display EC default schedule
The "Recommended Settings" button now changes the schedule mode to match the EC default schedule. If this mode would be different than the currently selected one, the user may choose to cancel (just as if they had clicked the radio button). If the user accepts the mode change, the EC default schedule settings will be applied. If the user cancels, no changes are made.