YUK-9597 Can't receive FDR data values through the RDEX interface.
The SQL select for an individual point reload was missing a condition in the where clause.  Our pointid from the point table didn't match the pointid from the fdrtranslation table.  Fixed condition - problem went away.  Verified the fix was the way it used to be in the roguewave SQL generated statements.