YUK-9423 DeviceAddressRange for MCT410fL needs to be updated
Primary change is to enable storing string values with some PaoTags. This changes the syntax for declaring PaoTags in the paoDefinition.xml file and changes how the file is processed. Going forward, a PaoTag can declare that it supports a value by specifying the type of the value (same syntax as YukonRoleProperty enum). Existing methods for checking if a PaoType supports a PaoTag work the same, but if the PaoTag is supported by the PaoType, its value can be read as either a String or a long.

Additional changes:
 - Changed paoDefinition.xsd to enforce uniqueness in several places to prevent errors.
 - Changed PaoStore to use Lists instead of Arrays and to better handle nulls.
 - Updated a lot of PaoDefinitionDaoImpl to use Multimaps and BiMaps where appropriate.
 - Removed a deprecated method on PointTemplate by updating its only callers (was only called from test code).
 - Add JUnit test code for PaoDefinitionDao methods where it was missing.