YUK-9716 Past Profile Collection errorMsg is not displaying
Added a hideReveal tag to display the error message supplied by the ProfileWidget.java controller if there is one.
Added an error message in the case that e-mail is left blank.
Removed the call to DateUtils.addDays(stopDate, 1) to prevent the user from requesting a past profile that begins today and ends yesterday.