YUK-9802 Detect/Handle &nbsp; in master.cfg.
Changed from replacing all non-ascii characters to replacing all space characters with the normal space character (e.g. &nbsp). Also will detect when any characters which are not printable characters or other standard chars (tab, newline, etc.) and print the line number and a warning. If a bad character is in the file, the system may still cause problems but will log an appropriate error.