YUK-9834 Stars Super Admin User > User/Group editor > User Permissions page displaying 'Failed Login Info'
Part of this issue seems to be fixed in 5.3.2 where the permissions link on the left doesn't show up when the user doesn't have the Administrator>LM User Assign role set to true. (YUK-9835)

I added the CheckRoleProperty annotation to the controller so if the user navigates directly to the URL, it will show the user not Authorized page.