YUK-9819 Stars operator > Hardware - error message from Add Thermostat popup cannot be seen - behind popup
Added a call to adjust the position of the serial number entry popup when clicked & on page load, moving it out of the way of the error message.  In cases where the bottom message still happens to be covered, the FlashScope messages now appear on the popup as well as the page so the user will always be able to see it.
Also added three wrappers to checkSerialNumber() in order to have a separate commandName & RequestMapping for each serial number popup box.  This prevents the three popups from sharing BindingResult error messages, though FlashScopes are still shared.  
Lastly, made each section of the page hidden when the account has no devices of that type available to add.