YUK-9883 DB Editor: Point Editor is missing fields to set decimal places and meter digits
Changed the frame size back to the original value of (435,600). Moved the state group label and dropdown to below the Unit Of Measure fields and moved the Decimal Places fields and Meter Digits to the lines below it, giving each of these their own lines. Also, I changed the minimum size of the Decimal and Meter Digit Spinners.