YUK-9887 Add printout on one of the web pages to indicate the current user's Energy Company
Added the current energy company name (if one is linked to the user) and the current username to the footer.

Changed from one line to 3 lines having The copyright on the first line, the yukon version, generated timestamp, energycompany name, and username on the second line. The third line will contain module name, page name and build info, if the roleproperties are set.