YUK-9778 Daily Usage Report dates not validated/Error messages not displayed
Added date validation to the Daily Usage Report section of the meter profile widget- previously there was no validation.
It now uses widgetActionRefresh to display validation errors to the user.
If the dates entered are valid, the user is redirected to the report.
The method of redirection has also been modified and now takes advantage of Yukon.ui.blockpage().