YUK-10007 IVVC bus should go into suspend mode when an LTC/Reg goes into auto mode.
Added code to detect a regulator in auto mode to the code for determining comms lost.  Now a regulator in auto mode will cause the system to go to a comms lost state, (stop sending heartbeats to let go of the regulators and cbcs - after appropriate retries) instead of sitting in a suspend mode and maintaining control of the cbcs and regulators.