YUK-10014 Trend widget on metering page the graph is shrunken on firefox and IE
Set a fixed height of 250px and a variable width of 100%. 
Moved the size specification from the div surrounding the graph to the trend tag itself. 
The graph now stays at the same height in all three browsers, and grows to fill its horizontal space, because redraw has been set to true in the trendSettings_* files.