YUK-9995 Smart Polling YUK-9996 More than one gateway does not update YUK-10087 improve retrieval of notifications YUK-10074 status showing connected but decommissioned
Fixed Xpath syntax to update all gateways
Threaded off gateway polling to increase rate of notification download
Changed updates to use timestamp in message to prevent the wrong state from showing.
Added ability to 'smart poll' When an action is performed on a device we will poll it for a CPARM amount of time.