YUK-6174 MultiSpeak - use Attributes for reading point data
Fix to how the i18n message scope is set up for Widgets. Because the postHandle is not guaranteed to be called and the afterCompletion is, there is a situation that can occur where we would pop off a scope even though we never pushed one on. This had the effect of breaking all other i18n lookups on the page.