YUK-10371 Inventory breadcrumbs on operator page are broken
Added module_config entries for thermostatHistory.single and thermostatHistory.multiple, as well as thermostatManual.single and thermostatManual.multiple pages.

Modified standardPageTags to reflect these module_config entries.

Changed OperatorThermostatHistoryController.java and OperatorThermostatManualController.java to add an additional attribute, pageNameSuffix, to the model.  It indicates whether the page is displaying a single thermostat or multiple thermostats, and points the view to the right module_config entry.