YUK-10418 IVVC Capcontrol server crashes silently on Tech Support's test/demo system.
Added a point request 'RequiredType' and assigned the bus watt and var points to it.  In the stale data check code set the ratio to 1.0 (all points required).  Check the iterator obtained from map::find() to make sure it isn't end() - supply sane default if it is.  Note: now that the required points are truly required, we should never get end() as an iterator for the bus watt and var points.