YUK-10249 Manual Commander > duplicate drop down option
Added semicolons to SQL Server 5.4_0 update.
Without them all the alter statements are processed as a group, if one fails they all fail together in the valids files, but commands that succeeded in the DB will not be rolled back (not good since the valids file then tries to re-do the successful commands).