YUK-9417 Lost functionality - Enable/Disable Program(s) from TDC feature is not available on web Demand Response Interface
Added "enable programs" and "disable programs" links to the actions sections of the control area detail and scenario detail pages. These open a confirmation popup that lets the user select programs within the parent object to enable or disable. In the case of disabling, the user may also choose to suppress restoration commands.