YUK-10329 Cap Control : Assigned Cap Bank Devices shows up in a Random order in Zone Editor window as its selected YUK-10666 Zone creation cap bank picker displays "Loading..." forever
Added "noBlockOnAdd" parameter to dynamicTable tag to prevent blocking the page when there are additional steps to carry out after clicking the add button, before adding the row.

Added an "addItems" method to dynamicTable that ensures item rows are added in order and refactored the zone details wizard page to use it.