YUK-10635 AMR - EIM Web Services - add RF support and verification to demand reset web service
Fixed the following bugs:
Currently the service is returning more than one lookupError element if there is more than one type with lookup errors.
More than one device group lookup failure results in an exception in the server.
Some of the tags in the output are missing the correct namespace.