YUK-10810 IVVC DispatchPointDataRequest::updateRejectedValues is not removing all non-updated values from _values map, so the non-updated values are being used in Analysis.
Reworked the processNewMessage function to properly sort the incoming point data messages when they come in.  Now we keep the newest 'good' message.  It used to keep the newest message regardless of quality, then filter qualities when _complete was set (which wasn't happening with missing data).
Added new unit testing to capture behavior.