YUK-10768 password and authType columns not always set appropriately
Merged "set password" and "change password" concepts which did the same thing but were inconsistently used.
Removed unused password recovery.
Fixed password change to update authentication type to use the current authentication type instead of sticking with the one used on the account.
When a login is created, use the AuthenticationService to properly update the password rather than storing it as clear text even if the authentication type was not plain.