YUK-10825 Webservice Schema validation failures
Changed RunThermostatProgramEndpoint.java so that the success element is added within the try block, removing the possibility for both failure and success elements to be returned together in the same response.

Made the complex type SimpleSuccessOrFailure use the element name 'failure' instead of 'error'.  It is now in line with the actual response being sent by all 3 web service responses that make use of the SimpleSuccessOrFailure type.
This has fixed schema validation issues with these three web service responses:

SendManualThermostatSettingResponse
RunThermostatProgramResponse
ForgotPasswordResponse