YUK-11046 CBC 8000 Firmware Version is displaying incorrectly in the 4-tier timestamp information window
Changed the way the cbc7/8000s format the firmware version and send to the client for display.  Now it is a 6 bit char, 8 char max message embedded in the mantissa of the firmware revision point data message value.  Updated client code to build this message.  Added unit testing for the 7000 and updated the existing tests for the 8000