YUK-10639 Improve dropdown for Attributes and limit list of Attributes appropriately
Added attribute groups as Lists in BuiltInAttribute.java.  The result is maps of group names to lists of attributes.
Added a new attribute, groupItems, to the selectWithItems tag.  When set to true, <optgroup> tags are rendered within a dropdown list to display attributes in groups, with each group having a heading.
Converted the ArchivedValuesExporterController to use the groupItems functionality of the selectWithItems tag.
Did the same conversion on ArchiveDataAnalysis' home.jsp
Also added a new sort method to the ObjectFormattingService to sort the group headings first, and then the individual entries within each group.