YUK-11154 DB CHANGE - Adding a no password policy
Added a new YukonGroup with the group name 'Password Polciy Grp2'.
Added password policy role properties from the new group.
The role properties configure the password policy so that passwords do not expire, are not expired by default, and have no requirements for password complexity.
Any users who are not already a member of a group that specifies the password policy role will be added to the new password policy group.