YUK-10363 SelectCom LCR support using native RF radio
Now subtracting an hour for each interval data reading.  This is because the startTime attribute encodes the most recent reading time, not the first in the list.
Also reversed the parsing of the runtime and shed time bytes from each intervalData element (runtime is the most significant byte, shed time the LSB).