YUK-11218 Add support for master.cfg encrypted values to C++ applications YUKRV-4904
Review changes.  Added unit testing to capture and document behavior of the hex string to byte array function.  Added some comments to clarify the unnamed namespace.  Put the encryption stuff inside the Cti namespace.  Patched the encryption unit test that is currently failing on the production build machine -- report an error but NOT fail the unit test.