YUK-11371 Change Password issues
Added statements to the 5.5.3 update script which raises the minimum password length and policy quality check role property default values from 0 to 1.  It also updates any existing role groups, but only when the value is still set to 0.
Anyone who has already raised their minimums won't have them reset to 1.

These values were set to 0 in the 5.5.2 update script, which was technically allowing a blank password to be 'legal'.