YUK-11406 IVVC Enhancement: Allow Multi-Tap LTC/Regulator Operations for OverVoltages
Added multi-tap capabilities to the bus analysis stage of the IVVC algorithm.  See the PDF attached to the issue for further clarification.  Developed estimated voltages based on multiple taps down.  Calculated a bus weight for this situation and if it is more optimal than switching a cap bank then we execute the multi-tap solution.  Multi-tap only occurs if we have an over-voltage somewhere on the subbus.