YUK-11553 z_meter_mct_base_desig - change default to stars
Changed the default value of the 'z_meter_mct_base_desig' role property from 'yukon' to 'stars' in the creation scripts as well as the update scripts.

In the update scripts, a warning will be issued if "yukon" meters exist in a STARS system. If they do the user will be notified that they have to manually set z_meter_mct_base_desig to the correct value after the upgrade is complete.