YUK-11587 Recreating the CBCConfiguration2_View to fix a datatype issue in oracle
Added an error ignore annotation to the DROP VIEW statement.  

One customer database did not have this view when the update was executed.  Since precedent exists we should guard against blocking the update for customers who don't have this view.
Instead of stopping the update entirely we will ignore the DROP VIEW error and continue on to create the view.