YUK-11745 The Yukon log should not print any master.cfg parameters that are encrypted
Changed encrypted values from master config to stay encrypted in memory. These will be decrypted on demand when the values are requested. This helps prevent logging these sensitive values

Changed CryptoException to be RuntimeException instead of Exception. This seems appropriate since we are not able to recover from this in most cases.