YUK-11762 CYME Integration does not accept float values for %
Changed to accept float values from the 24 hour simulation file.
Cleaned up some of the simulation code.
Exceptions will now print out that the simulation was halted.
The point will now reset on exceptions so we we can see on the IVVC pages that the 24 simulation stopped running.
Added the concept of Time to the load simulations. We will now start at the appropriate time of day in the file instead of starting at the beginning as "now"