YUK-11482 add missing log directories to support bundle
Adding two new log directories to the Support Bundle
1.	<yukon base>\Server\web\logs
2.	<yukon base>\Client\Log

For the Tomcat log files, the ability to filter based on lastModified date on the File and also the date in the filename was added.

The default is that the Tomcat logs are required, but the Client\Log files are optional.

With the addition of the Client logs, there are two peer directories, Client and Server. The structure of the Zip file will be updated to reflect there are two peer log directories: