YUK-11961 Divide by zero in GroupCommandCompletionAlert: If no commands complete, the percent success message displays '?%'
GroupCommandCompletionAlert.java will now display an appropriate message if the completed count is 0, instead of displaying the result of a divide by zero error.
GroupMeterReadController.java also benefited from a simple change that will display 0% instead of ?% for a similar divide by zero error on the group meter read page.