YUK-11912 **** ERROR **** DB EXCEPTION 23000 Violation of PRIMARY KEY constraint 'PK_DYNAMICCCMONITORBANKHISTORY'
Added code to aid in the debugging of this issue.  It is only compiled in in debug mode and is hidden behind a temporary CParm (CAP_CONTROL_MONITOR_POINT_DEBUGGING) which defaults to false.  The debugging code tracks creation and deletion of monitor points, dumps their pointID, deviceID and a stack trace of where they were created from.