YUK-12187 **** ERROR **** DB EXCEPTION 23000 Violation of PRIMARY KEY constraint 'PK_DYNAMICCCMONITORBANKHISTORY' (2)
Initial checkin.  Removed some unused code and did a general cleanup of the monitor point class.  Refactored the SQL in the bus store to load the static and dynamic data at object construction.  Removed setting the _dirty flag on static data (it is only used to sync dynamic data to the database).