YUK-12742 Disable AD SSL and LDAP SSL for existing customers
Added insert statements to 6.0.4 update scripts which will insert new global settings with the names 'AD_SSL_ENABLED' and 'LDAP_SSL_ENABLED' and the value '0' for both settings. This disabled these SSL settings for existing customers. 

No changes are required to the update scripts since they will not specify the setting in the database and so the default setting value '1' will be pulled from the enum and used for new installations.