YUK-12754 Stars - Deleting an account displayed 'error occurred' page +review @b1afc @f84b5 @cd52e @ma7212
Modified four foreign key constraints adding the ON DELETE CASCADE option to:
FK_AcctThermSch_CustAcct
FK_AccThermSchEnt_AccThermSch
FK_InvToAccThermSch_AccThermSc
FK_ThermEventHist_AcctThermSch

Now when deleting a thermostat schedule (or the account that it is on) all of the schedule entries and thermostat history events are guaranteed to be cleaned up.  There will not be an exception due to a foreign key reference between account and archived account thermostat schedules.