YUK-11837 Area level does not display VRD as it used to
Added a check for the children of the child substation being in VRD mode.  The bug was that we were only checking if the substation VRD flag was true, which it was not... The substations childVRD flag was because one of the substation buses was in VRD mode instead.

Simplified CtiCCArea::checkAndUpdateChildVoltReductionFlags() a bit.