YUK-12980 Add logging of LastCommunication times to LCR3102
Fixed a couple of bugs found during testing.  The pre-population of the row wasn't working because the device id wasn't properly being set in the table objects DecodeDatabaseReader function.  The timestamps were being updated on invalid data because invalid data has a non-zero value.  Added a check of the quality of the data to only update if non-zero and normal quality.