YUK-13608 Consistently archive Diversified load and Maximum kw points
Added blocking versions of getProgramValue() which wait for the computation to complete.
These new methods are now used by RepeatingEstimatedLoadTask which fixes the problem where values that were not immediately available in the cache were not being written to the database.
Removed the circular spinner and replaced with the standard yellow flash whenever a value is updated on a web page.