YUK-13804 Add column to support grouping rescheduled jobs together
Added new column, JobGroupId to the tables Job and FileExportHistory.
Converted existing JobId values in the Job table to the Job.JobGroupId column and parsed through them to find the max id value for a given BeanName. This is then used as the key when matching JobGroupId from the Job table to the FileExportHistory.JobGroupId column.