YUK-13904 Adding or Removing a CapBank to Time of Day Substation Causes Error Message in CapControl Log
The calculateKVARSolution function has a path through it for every control unit type.  The error message was added to make sure we didn't forget to update this function for any new control unit type, which we conveniently forgot to update for time of day control.

Added a path for time of day control.