YUK-13015 Support LTC and Regulator control through set point changes
Initial check-in for set point control.  Currently not accessible yet due to lack of device configuration support.  Added support for set point commands to the regulators, and adapted the algorithm to compute voltage changes instead of tap position changes.  The regulator is responsible for figuring out how to adjust it's voltage now.