YUK-14191 Alter RPHTag Table structure.
Removed the TagName column and replaced it with six bit fields.  The fields indicate which of the tags applies to a particular ChangeId.  Duplicate ChangeId values are no longer allowed in the table with multiple different TagName values.  Now each ChangeId has exactly one row that describes all of its tags.