YUK-12443 Extract point_info.freeze_bit
Moved the freeze_bit into a new struct, frozen_point_info and changed the mct4xxx decoders to use that new struct.

Added a test_dev_mct420_getDemandData to test_dev_mct410.cpp to match the mct410 test.  There appears to be differences between the 410 and 420 when decoding DemandData that I can't account for.