YUK-14462 Estimate time to convert existing C++ makefiles into Visual Studio projects
This change covers an issue with logManager instantiation and the fact that C++ global initialization have no defined order of execution between modules.  The change puts the instantiation of the log managers and the initialization of the pointers to the 2 loggers in the same module where order is guaranteed.  The issue was uncovered when being built with the new Visual Studio Projects.