YUK-14581 Removal of "Enroll Multiple Programs per Category" from Consumer Info
Transferred any existing role property values for the 'Enroll Multiple Programs Per Category' role property to the new energy company setting of the same name.  Values were taken from the role group that was associated with the energy company admin operator user.  If no value is found associated with that user, an error is raised during the update.