YUK-15144 Refactor more Capcontrol C++ code without breaking anything
Added more to the unit test to document the behavior of most of the trivial setters and getters with regards to what they do to the dirty _dirty flag and the _areaUpdatedFlag.  Added the framework around which to refactor the dynamic data portion of the areas, it is not yet incorporated.