YUK-15260 Continue C++ substation refactoring without breaking anything YUKRV-9545
Integrated season and holiday managers into CapControl and modified area scheduling semantics to take advantage of them.  Added support for DB change messages for the schedule managers.

Changed loading of areas to do all the heavy lifting into a local data structure and then when finished, lock the mux and exchange the existing for the new.