YUK-15296 Refactor SubstationBus C++ without breaking anything
Removed the custom implementation of the copy constructor and assignment operator and allowed the compiler to create as directed (a default copy constructor and no assignment operator).  Reworked the replicate function to do a shallow copy via the (now defaulted copy constructor) and then to a deep copy on its contained feeder collection (which had been done in the (now deleted) assignment operator).