YUK-16682 Integrate CBC heartbeat into CapControl server
Added code to the control loop to send or stop CBC heartbeat messages as appropriate except in the case of an IVVC strategy, that is handled in the IVVC algorithm.  In that case, we start and stop heartbeats at the same time as we start and stop the regulator heartbeats.