YUK-16650 Porter should send each port's queue size as a point update YUKRV-11223
Added Port type -> device type mappings in resource_helper.cpp.

Defined a struct to hold a multi message of point update messages and the total queue count for all ports in porter.cpp. That is the struct that is used to accumulate and eventually send the port queue sizes as one multi message.