YUK-17771 Continuous Build site crawler seeing illegal use of attribute error for PORT_QUEUE_COUNT point YUKRV-11777
The issue with the site crawler was caused by ports that were created without their default points. I updated two spots in Yukon where the default points were not being created for ports. To find the spots in code that needed to be changed I searched Yukon for the method createPort().