YUK-17596 Compare error codes that porter returns to the error codes in error-code.xml for device config. YUKRV-11795
Review changes for the porter error codes test which makes sure error codes have been defined in all of the right places.
Most changes were for cleaning up code and improving efficiency.
Error code 37 was re-added to client as an error code so we know when porter has sent an error code of 37.