YUK-17904 Don't send the AutoBlockEnable command for SetPoint configured CL-6B/7 regulators YUKRV-11944
Modified the IncrementingKeepAlivePolicy to have two different implementations. One of them is for when the regulator is in SetPoint control mode and therefore does not send the AutoBlockEnable command inside of the SendKeepAlive() method.

Created a unit test to make sure that a regulator with the Incrementing policy and SetPoint control does not send the AutoBlockEnable command.