YUK-17904 Don't send the AutoBlockEnable command for SetPoint configured CL-6B/7 regulators YUKRV-11944
Modified the IncrementingKeepAlivePolicy to have two different implementations. One of them is for when the regulator is in SetPoint control mode and therefore does not send the AutoBlockEnable command inside of the SendKeepAlive() method.

Modified IncrementingKeepAlivePolicy to not have multiple implementations but
instead take an AutoBlock enum as a parameter to determine whether to send or 
suppress the AutoBlockEnable command.