YUK-17992 Infrastructure warnings event logs should update only when condition first appears YUKRV-11983
Implemented the same logic that the smart notifications use for creating infrastructure warnings notifications: do not add an event to the event log if the current warning was in the previously calculated list of warnings.

This logic uses the .equals method in the Infrastructure warnings method which doesn't check the args for the warning. I am not sure if it should or not.