YUK-18022 Substation Bus and Feeder Voltages Not Logged to the DmvMeasurementData Table During DMV Test YUKRV-12031
Review changes.  Changed the way the bus and feeder watt and var points were handled in the point request.  The predefined category also had voltage points in it which were being ignored.  Now I log watt and var point IDs in a collection and use that to gather and process the feasibility data.