YUK-18006 CBC Detail Doesn't Show Config Name if Heartbeat Config Category Missing YUKRV-12056
Added alert messages when categories are required and (none) messages when categories are not required to both VIEW and EDIT displays for the cbc detail page. This also involved some additions to the javascript for the cbc detail page that updates the messages for categories when the pickers are used. Also added a method to the PaoDefinitionDao to get all required categories for a given paoType.