YUK-18516 - Make Note deletion a "real" delete YUKRV-12405
Modified the pao note service and the pao note dao to delete the PaoNote from the database rather than update its status to 'D'.

Removed the YukonUser arguments from the delete methods as well as the references to those methods since the user is not longer needed.

Removed 'D' status from the PaoNoteStatus enumeration.

Removed checks in the PaoNotesDao that filter out notes with a D status.