YUK-18414 - Notes UI - Notes Popup YUKRV-12403
Changed request method for note popup note deletion to RequestMethod.DELETE and made the necessary jsp changes as well.

Reverted the change made for the notes search link of the site map page.

Removed the global paoId variable from the notes popup javascript.

Used ajaxSubmit for successful form submit when creating a note in the popup.