YUK-18414 - Notes UI - Notes Popup YUKRV-12403
Changed the way the model is set up for the notes popup so that the only information needed in the controller is the paoId for the desired notes popup.

Fixed name collisions between the notes popup and notes widget.

Fixed issue with a new popup being created every time the notes widget renders when the desired outcome is to re-use the same widget.

Modified the link to the search page that is in the upper right hand corner of the notes popup to contain the correct key.