YUK-18414 - Notes UI - Notes Popup YUKRV-12403
Fixed the word-wrapping issue in the notes widget, popup, and search page so that they are using word-wrap: break-word instead of breaking all (breaking in the middle of words).

Added a style option to the sort tag.

Modified the paoNotesPopup to use colspan 2 on the notetext field when the user is not editing the note (editing buttons hidden) so that the table doesn't have a weird gap in it.

Reduced the number of characters for the note preview in the meter information widget to 28 to be safe about wrapping.