YUK-17683 - Gateway Detail > Map Network - add details popup on map pin click YUKRV-12510
Pulled common mapping code out from the gateway details page and the map network page and placed it in a new locationContainer.jsp. 

Removed the javascript that was being used by the gateway detail page and included the map network page javascript instead.

This means the jsp code and javascript code for mapping on the gateway detail page and on the map network page are now in one place, however I did have to modify the GatewayDetailController.