YUK-17190 Fix CapControl point registration fragility
Removed the old point registration code, built a new point registry tracking object and integrated it into CapControl.  We will now track the number of instances a point ID is registered with dispatch outside the connection itself and register and unregister the point ID as necessary.  We also are watching the state of the dispatch connection to monitor when we need to do a full point re-registration.