YUK-18741 - Add Notes widget to hardware detail page YUKRV-12623
Changed the notes widget from using td width=x% for formatting to yukon's column-x-x formatting to allow for dynamic sizing of the widget so that it looks good on the hardware detail page.

Updated the create new note area on the pao notes popup to use the dynamic yukon column-x-x formatting as well.

Fixed a bug in the Notes widget controller that would un-hide the border lines when a note was created/edited/deleted.