YUK-18632 Allow user to view Device Notes from mapping popups YUKRV-12638
Added notes icon next to gear dropdown in the mapping popup on both the Collection Actions Map Devices page and the Map Network page.

Removed the gear dropdown from mapPopup.jsp because the the same gear dropdown is included in info.jsp which gets ajaxed in on the both the Collection Actions Map Devices page and the Map Network page.