YUK-18832 Some pin popups not displaying expected info YUKRV-12634
When Yukon asks Network Manager for RfnMetaData and the comm_status of the device is "Unknown" we will not get the primary gateway back. This commit adds an error, warning, or success class to the Status field in the mapping popup to call out why or why not the primary gateway field is populated or empty.

This is the same indication that the rfnDeviceMetaData widget gives the user and any change to the mapping popup or widget should be consistent between both places that we display this information.