YUK-18785 Hide notes icon after deleting all notes in notes popup YUKRV-12640
Added the cog dropdown back in to mapPopup.jsp under a new div with the notes icon so that the cog and notes icons can be displayed properly. Modified the common mapping javascript to interact with that new div.

Added a method to yukon.tools.paonotespopup.js and PaoNotesSearchController to determine if a device id has notes and then hide or show all of the icons for that device on the page.