YUK-19897 - Itron data collection task doesn't run when devices are created
Changed the service so that the task always runs periodically, but checks for Itron devices before doing any work. The task still won't run if the data collection interval is set to zero, but should pick up any changes and resume running if the interval is updated to a value >0. No restarts should be necessary.