YUK-20468 Delete old FDR debug mode code
Removed the code to create the old FDR debug mode thread code.  This code would send an incrementing value point data message to dispatch every second for every point in the send list.  The code was buggy and crashing FDR when enabled and running in conjunction with the Java client FDR emulator.