YUK-20701 DR Setup? - creating Control Area with control window, with Timed program causing log errors
If a program has invalid control windows it will set the start and stop times to -infinity and if constraints are checked the -infinity date will throw an invalid year exception during the check. The fix was to use the global invalid date that LoadManagement uses of 1/1/1990 instead because it parses properly.