YUK-22555 CapControl disabled point not synced properly on dbchange.
Backported changes from master to this branch, which is a patch for Duke East.  Also updated the versioning stuff to use the git stuff rather than the svn stuff.  Added a field for a custom patch name to the build_version.properties file and associated parsing in the setVersion script.