YUK-22875 - Data Export on Interval
Initial review of back-end code. Still needs UI, schedule support, cleanup, tests.

Dynamic Attribute Exports can be flagged as "on interval", with a specified interval. Any data within the range that is not on the interval will be excluded. Any intervals within the range that have no data will generate a data row (as long as the strategy for missing values is "leave blank" or "fixed value").